requirements.txt
streamlit
nltk
google-generativeai